Use GHC 9.0.2 and base 4.16.x
The Decode module now has branching logic to see which version of base
is being used. This allows us to use MonadFail for base >= 4.13

Additional stack files have been added to ensure the library works with
GHC 8.6.x and GHC 9.2.x

We also have updated to use the key types present in cryptonite, for
more flexiblity.